% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getting_data}
\alias{getting_data}
\title{Data getting}
\usage{
getting_data(groups, researchers)
}
\arguments{
\item{groups}{A dataframe with information about groups and links from GrupLAC}

\item{researchers}{A dataframe with ID from google scholar form researchers}
}
\description{
This function get information from GrupLac.
}
\details{
Extracts data from Minciencias web pages about research groups and researchers.
}
\examples{
\dontrun{
Load data in a dataframe of r
groups <- read.csv(".../groups_information.csv", header=T, sep=",")

margaret_data <- getting_data(groups)

#or just
margaret::getting_data(groups)
}
}
