% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{homologadas_publindex}
\alias{homologadas_publindex}
\title{"Publindex homologación de revistas" 2003 to 2022}
\format{
A dataframe with information about journal ranking from Publindex
}
\source{
\url{https://scienti.minciencias.gov.co/publindex/#/revistasHomologadas/buscador}
}
\usage{
homologadas_publindex
}
\description{
"Publindex homologación de revistas" 2003 to 2022
}
\keyword{datasets}
