% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{international_journals_2016_2020}
\alias{international_journals_2016_2020}
\title{Publindex homologación de revistas 2016 to 2020}
\format{
A dataframe with information about journal ranking from scimago
}
\source{
\url{https://scienti.minciencias.gov.co/publindex/#/revistasHomologadas/buscador}
}
\usage{
international_journals_2016_2020
}
\description{
Publindex homologación de revistas 2016 to 2020
}
\keyword{datasets}
