% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{international_journals_2022}
\alias{international_journals_2022}
\title{Publindex homologación de revistas 2022}
\format{
A dataframe with information about journal ranking from Publindex
}
\source{
\url{https://scienti.minciencias.gov.co/publindex/#/revistasHomologadas/buscador}
}
\usage{
international_journals_2022
}
\description{
Publindex homologación de revistas 2022
}
\keyword{datasets}
