% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{internationals_journal_2021}
\alias{internationals_journal_2021}
\title{Scimago 2021}
\format{
A dataframe whit information about journal ranking from scimago
}
\source{
\url{https://github.com/ikashnitsky/sjrdata} \url{https://www.scimagojr.com/}
}
\usage{
internationals_journal_2021
}
\description{
Scimago 2021
}
\keyword{datasets}
