% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{national_publindex}
\alias{national_publindex}
\title{Publindex Nacional de revistas 2022}
\format{
A dataframe with information about journal ranking from Publindex
}
\source{
\url{https://scienti.minciencias.gov.co/publindex/#/revistasPublindex/historico}
}
\usage{
national_publindex
}
\description{
Publindex Nacional de revistas 2022
}
\keyword{datasets}
