% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scimago_2020}
\alias{scimago_2020}
\title{Scimago 2020 data}
\format{
A dataframe with information about journal ranking from scimago
}
\source{
\url{https://github.com/ikashnitsky/sjrdata} \url{https://www.scimagojr.com/}
}
\usage{
scimago_2020
}
\description{
Scimago 2020 data
}
\keyword{datasets}
