% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scimago_categories}
\alias{scimago_categories}
\title{Scimago categories}
\format{
A dataframe with information about journal ranking from scimago
}
\source{
\url{https://github.com/ikashnitsky/sjrdata} \url{https://www.scimagojr.com/}
}
\usage{
scimago_categories
}
\description{
Scimago categories
}
\keyword{datasets}
