% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scimago_data}
\alias{scimago_data}
\title{Scimago 2016-2019 data}
\format{
A dataframe with information about journal ranking from scimago since 2016 to 2019
}
\source{
\url{https://github.com/ikashnitsky/sjrdata} \url{https://www.scimagojr.com/}
}
\usage{
scimago_data
}
\description{
Scimago 2016-2019 data
}
\keyword{datasets}
