% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getting_data}
\alias{getting_data}
\title{data_getting}
\usage{
getting_data(df, researchers)
}
\arguments{
\item{df}{A dataframe that has group's informations}

\item{researchers}{A dataframe with ID from google scholar form researchers}
}
\description{
data_getting
}
\details{
This function get information from GrupLac
}
