% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{international_journals_2016_2020}
\alias{international_journals_2016_2020}
\title{Scimago 2016 to 2020}
\format{
A dataframe whit information about journal ranking from scimago
}
\source{
\url{https://github.com/ikashnitsky/sjrdata} \url{https://www.scimagojr.com/}
}
\usage{
international_journals_2016_2020
}
\description{
Scimago 2016 to 2020
}
\keyword{datasets}
