% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{internationals_journal_2021}
\alias{internationals_journal_2021}
\title{Publindex homologación de revistas 2021}
\format{
A dataframe with information about journal ranking from scimago
}
\source{
\url{https://scienti.minciencias.gov.co/publindex/#/revistasHomologadas/buscador}
}
\usage{
internationals_journal_2021
}
\description{
Publindex homologación de revistas 2021
}
\keyword{datasets}
